utilisateur edit Security

+++ LA-Factory-boot/src/main/java/fr/formation/Controller/UtilisateurController.java
@@ -1,5 +1,8 @@
 package fr.formation.Controller;
 
+import javax.servlet.http.HttpSession;
+import javax.swing.JOptionPane;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -25,9 +28,15 @@ public class UtilisateurController {
 	}
 
 	@GetMapping("/ajoutUtilisateur")
-	public String get(Model model) {
-		model.addAttribute("utilisateur", new Utilisateur());
-		return "ajoutUtilisateur";
+	public String get(Model model, HttpSession session) {
+		Utilisateur user = daoUtil.findById((Integer) session.getAttribute("userId")).get();
+		if(user.getRole().equals("ADMIN")) {
+			model.addAttribute("utilisateur", new Utilisateur());
+			return "ajoutUtilisateur";
+		}
+		else {
+			return "listeUtilisateur";
+		}
 	}
 
 	@PostMapping("/ajoutUtilisateur")
@@ -40,24 +49,40 @@ public class UtilisateurController {
 	
 	
 	@PostMapping("/editUtilisateur")
-	public String saveedit(@ModelAttribute Utilisateur utilisateur) {
-
-		daoUtil.save(utilisateur);
-
-		return "redirect:listeUtilisateur";
+	public String saveedit(@ModelAttribute Utilisateur utilisateur, HttpSession session) {
+		Utilisateur user = daoUtil.findById((Integer) session.getAttribute("userId")).get();
+		if(user.getRole().equals("ADMIN") || user.getNom().equals((String) utilisateur.getNom()) ) {
+			daoUtil.save(utilisateur);
+			return "redirect:listeUtilisateur";
+		}
+		else {
+			return "redirect:listeUtilisateur";
+		}
 	}
 
 	@GetMapping("/supUtilisateur")
-	public String delete(@RequestParam int id) {
-		daoUtil.deleteById(id);
-
-		return "redirect:listeUtilisateur";
+	public String delete(@RequestParam int id, HttpSession session) {
+		Utilisateur user = daoUtil.findById((Integer) session.getAttribute("userId")).get();
+		if(user.getRole().equals("ADMIN")) {
+			daoUtil.deleteById(id);
+			return "redirect:listeUtilisateur";
+		}
+		else {
+			return "redirect:listeUtilisateur";
+		}
 	}
 
 	@GetMapping("/editUtilisateur")
-	public String update(@RequestParam int id, Model model) {
-		model.addAttribute("utilisateur", daoUtil.findById(id).get());
-		return "ajoutUtilisateur";
+	public String update(@RequestParam int id, Model model, HttpSession session) {
+		Utilisateur changeUser = daoUtil.findById(id).get();
+		Utilisateur user = daoUtil.findById((Integer) session.getAttribute("userId")).get();
+		if(user.getRole().equals("ADMIN") || user.getNom().equals((String) changeUser.getNom()) ) {
+			model.addAttribute("utilisateur", daoUtil.findById(id).get());
+			return "ajoutUtilisateur";
+		}
+		else {
+			return "redirect:listeUtilisateur";
+		}
 	}
 
 }
